Show metrics as Splunk results and add noop option
This change sends the metrics and values back to Splunk so they can be
displayed to the user. It also adds a "--noop" option to the graphite
command which shows the results but doesn't send them to the Graphite
host. This option is useful for testing to make sure the metric names
look correct before commiting them to Graphite.